+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Reworked how backends (asdf, etc) are configured. Instead of being prefixed on the version, they are now on the tool identifier. Example: `zig = "asdf:0.14"` -> `"asdf:zig" = "0.14"`
   - Unfortunately, there was no simple way to support both patterns in parallel, so this is a hard break.
   - This opens the door for new functionality in the future, and also makes more logical sense.
+- Removed the `[tools.*.backend]` setting, as it doesn't work with this new pattern.
 - **WASM API**
   - Renamed `ToolContext` to `PluginContext`.
   - Renamed `ToolUnresolvedContext` to `PluginUnresolvedContext`.